requirements.txt
numpy
pygame
pyyaml